+++ Station.cpp
@@ -4,10 +4,10 @@ Station::Station() {
 
 	this->longitude = 0.0;
 	this->latitude = 0.0;
-	this->value = 0;
+	this->value = 0.0;
 }
 
-Station::Station(float longitude, float latitude, int value) {
+Station::Station(float longitude, float latitude, float value) {
 	
 	this->longitude = longitude;
 	this->latitude = latitude;
